refactor: improve skelton naming and design
