+++ src/components/Canary/index.tsx
@@ -28,7 +28,7 @@ import { useHealthPageContext } from "../../context/HealthPageContext";
 import { isCanaryUI } from "../../context/Environment";
 import clsx from "clsx";
 import dayjs from "dayjs";
-import RowsSkeltonLoader from "../RowsSkeltonLoader";
+import HealthPageSkeltonLoader from "../SkeltonLoader/HealthPageSkeltonLoader";
 
 const FilterKeyToLabelMap = {
   environment: "Environment",
@@ -176,7 +176,7 @@ export function Canary({
   }, 400);
 
   if (isLoading && !checks.length) {
-    return <RowsSkeltonLoader showSidebar />;
+    return <HealthPageSkeltonLoader showSidebar />;
   }
 
   return (
